refactor: rename repo methods for readability
